+++ app/src/main/java/com/example/tutorsearcher/User.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 public abstract class User
 {
     private String name;
-    private int age;
+    private long age;
     private String gender; //male, female, other
     private String profilePic; //url to the profile pic
     private String email; //should be a usc email
@@ -29,7 +29,7 @@ public abstract class User
     {
         return name;
     }
-    public int getAge()
+    public long getAge()
     {
         return age;
     }
@@ -60,7 +60,7 @@ public abstract class User
     {
         return null;
     }
-    public int getNumRatings() {return -1;}
+    public long getNumRatings() {return -1;}
     public double getRating() {return -1;}
 
     //setters
@@ -69,7 +69,7 @@ public abstract class User
         profilePic = newPic;
     }
     public void setName(String newName){ name = newName; } // *** added setter for name
-    public void setAge(int newAge){ age = newAge; }
+    public void setAge(long newAge){ age = newAge; }
     public void setGender(String newGender){ gender = newGender; }
 
     // TUTOR ONLY BELOW... these will be overridden by polymorphism in Tutor class---
@@ -82,7 +82,7 @@ public abstract class User
     public void addCourses(String addCourse){
         throw new RuntimeException("This method is not supported except by an instance of a Tutor class");
     }
-    public void setNumRatings(int newNumRatings) {
+    public void setNumRatings(long newNumRatings) {
         throw new RuntimeException("This method is not supported except by an instance of a Tutor class");
     }
     public void setRating(double newRating) {
